Split git and report orchestration hotspots
