+++ changelog.d/unreleased/4179.internal.md
@@ -1,0 +1,21 @@
+---
+category: internal
+issues:
+  - 4179
+affected:
+  - src/CodeIndex/Cli/GitHelper.cs
+  - src/CodeIndex/Cli/GitProcessRunner.cs
+  - src/CodeIndex/Cli/ReportBundleWriter.cs
+  - src/CodeIndex/Cli/ReportCommandRunner.cs
+  - src/CodeIndex/Cli/ReportLogTailBuilder.cs
+  - tests/CodeIndex.Tests/GitProcessRunnerTests.cs
+  - tests/CodeIndex.Tests/ReportLogTailBuilderTests.cs
+---
+
+## English
+
+- Split git process capture and report bundle/log-tail orchestration into focused internal helpers, with direct contract coverage for diagnostic redaction and log truncation (#4179).
+
+## 日本語
+
+- git process capture と report bundle/log-tail の orchestration を専用の内部ヘルパーへ分割し、diagnostic redaction と log truncation の contract を直接検証するテストを追加しました (#4179)。
